Moved business logic to recipe_logic.py Moved logic to recipe_logic.py Added documentation as per #9
